+++ src/components/AppStart/index.js
@@ -1,4 +1,4 @@
 // @flow
-export { triggerLoadCoreEpic, loadAppEpic, loadCoreFailedEpic } from './appStart.epics';
+export { triggerLoadCoreEpic, loadAppEpic } from './appStart.epics';
 export { appStartActionTypes } from './appStart.actions';
 export { AppStart } from './AppStart.component';