+++ src/core_modules/capture-core/components/Pages/EnrollmentEditEvent/EnrollmentEditEventPage.container.js
@@ -4,7 +4,7 @@ import React from 'react';
 import { useDispatch } from 'react-redux';
 import { useHistory } from 'react-router-dom';
 import { useEnrollmentEditEventPageMode } from 'capture-core/hooks';
-import { useCommonEnrollmentDomainData } from '../common/EnrollmentOverviewDomain';
+import { useCommonEnrollmentDomainData, showEnrollmentError } from '../common/EnrollmentOverviewDomain';
 import { useTeiDisplayName } from '../common/EnrollmentOverviewDomain/useTeiDisplayName';
 import { useProgramInfo } from '../../../hooks/useProgramInfo';
 import { pageStatuses } from './EnrollmentEditEventPage.constants';
@@ -16,11 +16,47 @@ import { deleteEnrollment } from '../Enrollment/EnrollmentPage.actions';
 import { buildEnrollmentsAsOptions } from '../../ScopeSelector';
 import { convertValue } from '../../../converters/clientToView';
 import { dataElementTypes } from '../../../metaData/DataElement';
+import { useEvent } from './hooks';
+import type { Props } from './EnrollmentEditEventPage.types';
+import { LoadingMaskForPage } from '../../LoadingMasks';
+
+const getEventDate = (event) => {
+    const eventDataConvertValue = convertValue(event?.occurredAt || event?.scheduledAt, dataElementTypes.DATETIME);
+    const eventDate = eventDataConvertValue ? eventDataConvertValue.toString() : '';
+    return eventDate;
+};
+
+const getPageStatus = ({ orgUnitId, enrollmentSite, teiDisplayName, trackedEntityName, programStage, event }) => {
+    if (orgUnitId) {
+        return enrollmentSite && teiDisplayName && trackedEntityName && programStage && event
+            ? pageStatuses.DEFAULT
+            : pageStatuses.MISSING_DATA;
+    }
+    return pageStatuses.WITHOUT_ORG_UNIT_SELECTED;
+};
 
 export const EnrollmentEditEventPage = () => {
+    const { orgUnitId, eventId } = useLocationQuery();
+    const { loading, event } = useEvent(eventId);
+    const { program: programId, programStage: stageId, trackedEntity: teiId, enrollment: enrollmentId } = event;
+
+    return !loading && eventId ? (
+        <EnrollmentEditEventPageWithContext
+            programId={programId}
+            stageId={stageId}
+            teiId={teiId}
+            enrollmentId={enrollmentId}
+            orgUnitId={orgUnitId}
+            eventId={eventId}
+
+        />
+    ) : <LoadingMaskForPage />;
+};
+
+const EnrollmentEditEventPageWithContext = ({ programId, stageId, teiId, enrollmentId, orgUnitId, eventId }: Props) => {
     const history = useHistory();
     const dispatch = useDispatch();
-    const { programId, stageId, teiId, enrollmentId, orgUnitId, eventId } = useLocationQuery();
+
     const { program } = useProgramInfo(programId);
     const programStage = [...program.stages?.values()].find(item => item.id === stageId);
     const hideWidgets = useHideWidgetByRuleLocations(program.programRules.concat(programStage?.programRules));
@@ -29,6 +65,7 @@ export const EnrollmentEditEventPage = () => {
         history.push(`/enrollment?${buildUrlQueryString({ orgUnitId, programId, teiId })}`);
         dispatch(deleteEnrollment({ enrollmentId }));
     };
+    const onEnrollmentError = message => dispatch(showEnrollmentError({ message }));
     const onAddNew = () => {
         history.push(`/new?${buildUrlQueryString({ programId, orgUnitId, teiId })}`);
     };
@@ -36,26 +73,28 @@ export const EnrollmentEditEventPage = () => {
         history.push(`/enrollment?${buildUrlQueryString({ enrollmentId })}`);
     };
 
-    const onGoBack = () => history.push(`/enrollment?${buildUrlQueryString({ orgUnitId, programId, teiId, enrollmentId })}`);
+    const onGoBack = () =>
+        history.push(`/enrollment?${buildUrlQueryString({ enrollmentId })}`);
     const enrollmentSite = useCommonEnrollmentDomainData(teiId, enrollmentId, programId).enrollment;
     const { teiDisplayName } = useTeiDisplayName(teiId, programId);
     // $FlowFixMe
     const trackedEntityName = program?.trackedEntityType?.name;
     const enrollmentsAsOptions = buildEnrollmentsAsOptions([enrollmentSite || {}], programId);
     const event = enrollmentSite?.events?.find(item => item.event === eventId);
-    const eventDataConvertValue = convertValue((event?.occurredAt || event?.scheduledAt), dataElementTypes.DATETIME);
-    const eventDate = eventDataConvertValue ? eventDataConvertValue.toString() : '';
+    const eventDate = getEventDate(event);
     const { currentPageMode, cancel } = useEnrollmentEditEventPageMode(event?.status);
     cancel && onCancel();
     const dataEntryKey = `singleEvent-${currentPageMode}`;
     const outputEffects = useWidgetDataFromStore(dataEntryKey);
 
-    let pageStatus = pageStatuses.MISSING_DATA;
-    if (orgUnitId) {
-        enrollmentSite && teiDisplayName && trackedEntityName && programStage && event
-            ? (pageStatus = pageStatuses.DEFAULT)
-            : (pageStatus = pageStatuses.MISSING_DATA);
-    } else pageStatus = pageStatuses.WITHOUT_ORG_UNIT_SELECTED;
+    const pageStatus = getPageStatus({
+        orgUnitId,
+        enrollmentSite,
+        teiDisplayName,
+        trackedEntityName,
+        programStage,
+        event,
+    });
 
     return (
         <EnrollmentEditEventPageComponent
@@ -75,6 +114,7 @@ export const EnrollmentEditEventPage = () => {
             onAddNew={onAddNew}
             orgUnitId={orgUnitId}
             eventDate={eventDate}
+            onEnrollmentError={onEnrollmentError}
             eventStatus={event?.status}
         />
     );