+++ src/core_modules/capture-core/components/Pages/Search/SearchPage.actions.js
@@ -21,6 +21,8 @@ export const searchPageActionTypes = {
     FALLBACK_SEARCH_COMPLETED: 'FallbackWillPushToSearchPageWithoutProgramSelected',
     ALL_SEARCH_RELATED_DATA_CLEAN: 'CleanSearchRelatedData',
     FALLBACK_SEARCH_RELATED_DATA_CLEAN: 'CleanFallbackSearchRelatedData',
+    SEARCH_PAGE_OPEN: 'SearchPage.SearchPageOpen',
+    NAVIGATE_TO_NEW_USER_PAGE: 'NavigateToNewUserPage',
 };
 
 export const navigateToMainPage = () => actionCreator(searchPageActionTypes.TO_MAIN_PAGE_NAVIGATE)();
@@ -86,3 +88,8 @@ export const cleanSearchRelatedData = () =>
 
 export const cleanFallbackRelatedData = () =>
     actionCreator(searchPageActionTypes.FALLBACK_SEARCH_RELATED_DATA_CLEAN)();
+
+export const openSearchPage = () => actionCreator(searchPageActionTypes.SEARCH_PAGE_OPEN)();
+
+export const navigateToNewUserPage = () =>
+    actionCreator(searchPageActionTypes.NAVIGATE_TO_NEW_USER_PAGE)();