+++ src/core_modules/capture-core/components/ScopeSelector/hooks/useResetProgramId.js
@@ -1,24 +1,31 @@
 // @flow
 import { useHistory, useLocation } from 'react-router-dom';
-import { getUrlQueries } from '../../../utils/url';
-import { buildUrlQueryString } from '../../../utils/routing';
+import { buildUrlQueryString, getLocationQuery } from '../../../utils/routing';
+import type { PageContext } from './types';
 
 export const useResetProgramId = () => {
     const history = useHistory();
     const { pathname } = useLocation();
 
     const resetProgramId = (pageToPush: string = pathname) => {
-        const { programId, ...restOfQueries } = getUrlQueries();
+        const { programId, ...restOfQueries } = getLocationQuery();
         history.push(`${pageToPush}?${buildUrlQueryString({ ...restOfQueries })}`);
     };
 
-    const resetProgramIdAndEnrollmentContext = (pageToPush: string = pathname) => {
-        const { programId, enrollmentId, stageId, eventId, ...restOfQueries } = getUrlQueries();
-        history.push(`${pageToPush}?${buildUrlQueryString({ ...restOfQueries, enrollmentId: 'AUTO' })}`);
+    const resetProgramIdAndEnrollmentContext = (pageToPush: string = pathname, pageContext: PageContext) => {
+        const {
+            programId,
+            enrollmentId,
+            stageId,
+            eventId,
+            teiId,
+            ...restOfQueries
+        } = getLocationQuery();
+        history.push(`${pageToPush}?${buildUrlQueryString({ ...restOfQueries, teiId: teiId ?? pageContext?.teiId, enrollmentId: 'AUTO' })}`);
     };
 
     const resetProgramIdAndTeiId = (pageToPush: string = pathname) => {
-        const { programId, teiId, ...restOfQueries } = getUrlQueries();
+        const { programId, teiId, ...restOfQueries } = getLocationQuery();
         history.push(`${pageToPush}?${buildUrlQueryString({ ...restOfQueries })}`);
     };