+++ src/core_modules/capture-core/components/ScopeSelector/hooks/useResetTeiId.js
@@ -1,15 +1,17 @@
 // @flow
 import { useHistory, useLocation } from 'react-router-dom';
-import { getUrlQueries } from '../../../utils/url';
-import { buildUrlQueryString } from '../../../utils/routing';
+import { buildUrlQueryString, useLocationQuery } from '../../../utils/routing';
+import type { PageContext } from './types';
 
 export const useResetTeiId = () => {
     const history = useHistory();
     const { pathname } = useLocation();
+    const { programId, orgUnitId } = useLocationQuery();
 
-    const resetTeiId = (pageToPush: string = pathname) => {
-        const { programId, orgUnitId } = getUrlQueries();
-        history.push(`${pageToPush}?${buildUrlQueryString({ programId, orgUnitId })}`);
+    const resetTeiId = (pageToPush: string = pathname, pageContext?: PageContext) => {
+        history.push(`${pageToPush}?${buildUrlQueryString({
+            programId: programId ?? pageContext?.programId, orgUnitId,
+        })}`);
     };
 
     return { resetTeiId };