+++ src/core_modules/capture-core/components/ScopeSelector/index.js
@@ -5,6 +5,7 @@ export {
     setOrgUnitFromScopeSelector,
     resetCategoryOptionFromScopeSelector,
     resetAllCategoryOptionsFromScopeSelector,
+    resetOrgUnitIdFromScopeSelector,
 } from './ScopeSelector.actions';
 export {
     useSetProgramId,