+++ src/core_modules/capture-core/components/TopBarActions/TopBarActions.container.js
@@ -1,15 +1,14 @@
 // @flow
 import React, { useState } from 'react';
 import { useDispatch } from 'react-redux';
-import { batchActions } from 'redux-batched-actions';
+import { useHistory } from 'react-router-dom';
 import i18n from '@dhis2/d2-i18n';
 import { ActionButtons } from './TopBarActions.component';
-import { openNewRegistrationPageFromScopeSelector, openSearchPageFromScopeSelector } from './TopBarActions.actions';
 import { resetAllCategoryOptionsFromScopeSelector } from '../ScopeSelector/ScopeSelector.actions';
-import { resetProgramIdBase } from '../ScopeSelector/QuickSelector/actions/QuickSelector.actions';
-import { useReset, useSetOrgUnitId } from '../ScopeSelector/hooks';
+import { useReset } from '../ScopeSelector/hooks';
 import { ConfirmDialog } from '../Dialogs/ConfirmDialog.component';
 import type { Props } from './TopBarActions.types';
+import { buildUrlQueryString } from '../../utils/routing';
 
 const defaultContext = {
     openStartAgainWarning: false,
@@ -30,7 +29,6 @@ export const TopBarActions = ({
     selectedProgramId,
     selectedOrgUnitId,
     isUserInteractionInProgress = false,
-    customActionsOnProgramIdReset = [],
 }: Props) => {
     const [context, setContext] = useState(defaultContext);
     const {
@@ -48,24 +46,45 @@ export const TopBarActions = ({
         openSearchPageWithoutProgramId;
 
     const dispatch = useDispatch();
+    const history = useHistory();
     const { reset } = useReset();
-    const { setOrgUnitId } = useSetOrgUnitId();
 
     const startAgain = () => {
         dispatch(resetAllCategoryOptionsFromScopeSelector());
         reset();
     };
-    const newRegistrationPage = () => dispatch(openNewRegistrationPageFromScopeSelector());
-    const newRegistrationPageWithoutProgramId = () => {
-        const actions = [resetProgramIdBase(), openNewRegistrationPageFromScopeSelector()];
-        dispatch(batchActions(actions));
-        setOrgUnitId(selectedOrgUnitId, 'new', false);
+    const newRegistrationPage = () => {
+        const queryArgs = {};
+        if (selectedOrgUnitId) {
+            queryArgs.orgUnitId = selectedOrgUnitId;
+        }
+        if (selectedProgramId) {
+            queryArgs.programId = selectedProgramId;
+        }
+
+        history.push(`new?${buildUrlQueryString(queryArgs)}`);
     };
-    const searchPage = () => dispatch(openSearchPageFromScopeSelector());
+
+    const newRegistrationPageWithoutProgramId = () => {
+        const queryArgs = selectedOrgUnitId ? { orgUnitId: selectedOrgUnitId } : {};
+        history.push(`new?${buildUrlQueryString(queryArgs)}`);
+    };
+
+    const searchPage = () => {
+        const queryArgs = {};
+        if (selectedOrgUnitId) {
+            queryArgs.orgUnitId = selectedOrgUnitId;
+        }
+        if (selectedProgramId) {
+            queryArgs.programId = selectedProgramId;
+        }
+
+        history.push(`search?${buildUrlQueryString(queryArgs)}`);
+    };
+
     const searchPageWithoutProgramId = () => {
-        const actions = [resetProgramIdBase(), openSearchPageFromScopeSelector(), ...customActionsOnProgramIdReset];
-        dispatch(batchActions(actions));
-        setOrgUnitId(selectedOrgUnitId, 'search', false);
+        const queryArgs = selectedOrgUnitId ? { orgUnitId: selectedOrgUnitId } : {};
+        history.push(`search?${buildUrlQueryString(queryArgs)}`);
     };
 
     const handleOpenStartAgainWarning = () => {