+++ src/core_modules/capture-core/components/WidgetEventSchedule/AccessVerification/index.js
@@ -1,0 +1,2 @@
+// @flow
+export { NoAccess } from './NoAccess.component';