+++ src/core_modules/capture-core/components/WidgetProfile/DataEntry/helpers/types.js
@@ -1,0 +1,6 @@
+// @flow
+
+export type Geometry = {
+    type: string,
+    coordinates: Array<Array<Array<number>>> | { latitude: number, longitude: number },
+};