+++ src/core_modules/capture-core/components/WidgetStagesAndEvents/Stages/Stage/StageDetail/StageDetail.component.js
@@ -19,6 +19,7 @@ import { colors,
 import { sortDataFromEvent } from './hooks/sortFuntions';
 import { useComputeDataFromEvent, useComputeHeaderColumn, formatRowForView } from './hooks/useEventList';
 import { DEFAULT_NUMBER_OF_ROW, SORT_DIRECTION } from './hooks/constants';
+import { getProgramAndStageForProgram } from '../../../../../metaData/helpers';
 import type { Props } from './stageDetail.types';
 
 
@@ -60,6 +61,7 @@ const StageDetailPlain = (props: Props) => {
         events,
         eventName,
         stageId,
+        programId,
         dataElements,
         hideDueDate = false,
         repeatable = false,
@@ -71,7 +73,8 @@ const StageDetailPlain = (props: Props) => {
         columnName: 'status',
         sortDirection: SORT_DIRECTION.DESC,
     };
-    const headerColumns = useComputeHeaderColumn(dataElements, hideDueDate);
+    const { stage } = getProgramAndStageForProgram(programId, stageId);
+    const headerColumns = useComputeHeaderColumn(dataElements, hideDueDate, stage?.stageForm);
     const { loading, value: dataSource, error } = useComputeDataFromEvent(dataElements, events);
 
 
@@ -132,8 +135,6 @@ const StageDetailPlain = (props: Props) => {
             .slice(0, displayedRowNumber)
             .map(row => formatRowForView(row, dataElements))
             .map((row: Object) => {
-                const dataTableProgramStage = events[0].programStage;
-
                 const cells = headerColumns.map(({ id }) => (
                     <Tooltip
                         key={`${id}-${row.id}`}
@@ -143,7 +144,7 @@ const StageDetailPlain = (props: Props) => {
                         {({ onMouseOver, onMouseOut, ref }) => (
                             <DataTableCell
                                 key={id}
-                                onClick={() => !row.pendingApiResponse && onEventClick(row.id, dataTableProgramStage)}
+                                onClick={() => !row.pendingApiResponse && onEventClick(row.id)}
                                 ref={(tableCell) => {
                                     if (tableCell && row.pendingApiResponse) {
                                         tableCell.onmouseover = onMouseOver;